Fix al_read_directory crash on 64-bit Windows.
(cherry picked from commit 191141666160ba02aa6755cef56584627d54c915)

+++ src/fshook_stdio.c
@@ -111,7 +111,7 @@ typedef struct
    struct dirent dd_dir;
 
    /* _findnext handle */
-   long dd_handle;
+   intptr_t dd_handle;
 
    /*
     * Status of search:
